+++ graylog2-plugin-interfaces/src/main/java/org/graylog2/plugin/inputs/transports/NettyTransport.java
@@ -82,7 +82,7 @@ public abstract class NettyTransport implements Transport {
         if (configuration.stringIsSet(CK_BIND_ADDRESS) && configuration.intIsSet(CK_PORT)) {
             this.socketAddress = new InetSocketAddress(
                     configuration.getString(CK_BIND_ADDRESS),
-                    (int) configuration.getInt(CK_PORT)
+                    configuration.getInt(CK_PORT)
             );
         } else {
             this.socketAddress = null;
